% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_packages.R
\name{list_contents}
\alias{list_contents}
\title{List the contents of the package}
\usage{
list_contents()

list_contents()
}
\description{
This function lists all R script files in the package's R directory.
}
